requirements.txt
requests==2.22.0
PyHamcrest==1.9.0
pytest==5.0.1
lxml==4.4.0
